Implemented handlers Also fixed some bugs in the PostgreSQL driver

+++ repository/model.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Subject struct {
-	ID          int
+	ID          int64
 	Name        string
 	Description string
 }
 
 type Event struct {
-	ID         int
-	SubjectID  int
+	ID         int64
+	SubjectID  int64
 	Ocurrences int
 	InsertTS   time.Time
 	LastUpdate time.Time
@@ -23,13 +23,13 @@ func (e *Event) GetDate() string {
 }
 
 type Repository interface {
-	InsertSubject(s Subject) (int, error)
-	GetSubject(id int) (Subject, error)
+	InsertSubject(s Subject) (int64, error)
+	GetSubject(id int64) (Subject, error)
 	UpdateSubject(s Subject) (int64, error)
-	DeleteSubject(id int) (int64, error)
+	DeleteSubject(id int64) (int64, error)
 
-	InsertEvent(e Event) (int, error)
-	GetEvent(id int) (Event, error)
+	InsertEvent(e Event) (int64, error)
+	GetEvent(id int64) (Event, error)
 	UpdateEvent(e Event) (int64, error)
-	DeleteEvent(id int) (int64, error)
+	DeleteEvent(id int64) (int64, error)
 }
